Move two transformers(memcpy and insert_cast) from framework to optimizer.
