+++ cmake/onnxruntime.cmake
@@ -49,6 +49,7 @@ target_link_libraries(onnxruntime PRIVATE
     ${onnxruntime_libs}
     ${PROVIDERS_CUDA}
     ${PROVIDERS_MKLDNN}
+    onnxruntime_optimizer
     onnxruntime_providers    
     onnxruntime_util
     ${onnxruntime_tvm_libs}
@@ -57,7 +58,6 @@ target_link_libraries(onnxruntime PRIVATE
     onnxruntime_graph
     onnxruntime_common
     onnxruntime_mlas
-    onnxruntime_optimizer
     debug ${onnxruntime_EXTERNAL_LIBRARIES_DEBUG} optimized ${onnxruntime_EXTERNAL_LIBRARIES})
 
 set_property(TARGET onnxruntime APPEND_STRING PROPERTY LINK_FLAGS ${ONNXRUNTIME_SO_LINK_FLAG})
