+++ cmake/onnxruntime_python.cmake
@@ -58,6 +58,7 @@ set(onnxruntime_pybind11_state_libs
     ${onnxruntime_libs}
     ${PROVIDERS_CUDA}
     ${PROVIDERS_MKLDNN}
+    onnxruntime_optimizer
     onnxruntime_providers
     onnxruntime_util
     ${onnxruntime_tvm_libs}
@@ -65,8 +66,7 @@ set(onnxruntime_pybind11_state_libs
     onnxruntime_util
     onnxruntime_graph
     onnxruntime_common
-    onnxruntime_mlas	
-    onnxruntime_optimizer
+    onnxruntime_mlas
 )
 
 set(onnxruntime_pybind11_state_dependencies
