+++ onnxruntime/core/providers/cuda/cuda_execution_provider.cc
@@ -3,7 +3,6 @@
 
 #include "cuda_common.h"
 #include "cuda_execution_provider.h"
-#include "core/framework/transformer_memcpy.h"
 #include "core/framework/memcpy.h"
 #include "cuda_fence.h"
 #include "cuda_allocator.h"
