+++ onnxruntime/core/providers/cuda/cuda_execution_provider.h
@@ -4,7 +4,7 @@
 #pragma once
 #include "cuda_pch.h"
 #include "core/platform/ort_mutex.h"
-#include "core/optimizer/graph_transformer.h"
+#include "core/graph/constants.h"
 #include "core/framework/allocatormgr.h"
 #include "core/framework/execution_provider.h"
 #include "shared_inc/cuda_utils.h"
