+++ onnxruntime/core/providers/mkldnn/mkldnn_execution_provider.h
@@ -9,9 +9,9 @@
 #include <memory.h>
 
 #include "core/platform/ort_mutex.h"
+#include "core/graph/constants.h"
 #include "core/framework/allocatormgr.h"
 #include "core/framework/execution_provider.h"
-#include "core/optimizer/graph_transformer.h"
 
 namespace mkldnn {
 struct memory;
