+++ onnxruntime/core/session/inference_session.cc
@@ -16,8 +16,6 @@
 #include "core/common/logging/logging.h"
 #include "core/common/task_thread_pool.h"
 #include "core/graph/graph_viewer.h"
-#include "core/optimizer/graph_transformer.h"
-#include "core/optimizer/graph_transformer_mgr.h"
 #include "core/graph/graph_utils.h"
 #include "core/graph/model.h"
 #include "core/framework/allocatormgr.h"
@@ -25,7 +23,6 @@
 #include "core/framework/environment.h"
 #include "core/framework/execution_frame.h"
 #include "core/framework/graph_partitioner.h"
-#include "core/framework/insert_cast_transformer.h"
 #include "core/framework/kernel_def_builder.h"
 #include "core/framework/kernel_registry.h"
 #include "core/framework/ml_value_patterns_planner.h"
@@ -37,8 +34,11 @@
 #include "core/framework/session_state_initializer.h"
 #include "core/framework/tensorprotoutils.h"
 #include "core/framework/tensorutils.h"
-#include "core/framework/transformer_memcpy.h"
 #include "core/framework/utils.h"
+#include "core/optimizer/graph_transformer.h"
+#include "core/optimizer/graph_transformer_mgr.h"
+#include "core/optimizer/insert_cast_transformer.h"
+#include "core/optimizer/transformer_memcpy.h"
 #include "core/platform/notification.h"
 #include "core/providers/cpu/cpu_execution_provider.h"
 #include "core/session/CustomOpsLoader.h"
